traces: reorganize layout components
