+++ web/src/features/traces/Trace.tsx
@@ -9,11 +9,11 @@ import { SearchProvider } from "@/src/features/traces/contexts/SearchContext";
 import { JsonExpansionProvider } from "@/src/features/traces/contexts/JsonExpansionContext";
 import { PlayheadProvider } from "@/src/features/traces/contexts/PlayheadContext";
 import { TraceGraphDataProvider } from "@/src/features/traces/contexts/TraceGraphDataContext";
-import { TraceLayoutMobile } from "@/src/features/traces/components/_layout/TraceLayoutMobile";
-import { TraceLayoutDesktop } from "@/src/features/traces/components/_layout/TraceLayoutDesktop";
-import { TracePanelNavigation } from "@/src/features/traces/components/_layout/TracePanelNavigation";
-import { TracePanelDetail } from "@/src/features/traces/components/_layout/TracePanelDetail";
-import { TracePanelNavigationLayoutDesktop } from "@/src/features/traces/components/_layout/TracePanelNavigationLayoutDesktop";
+import { TraceLayoutMobile } from "@/src/features/traces/components/TraceLayoutMobile";
+import { TraceLayoutDesktop } from "@/src/features/traces/components/TraceLayoutDesktop";
+import { TracePanelNavigation } from "@/src/features/traces/components/TracePanelNavigation";
+import { TracePanelDetail } from "@/src/features/traces/components/TracePanelDetail";
+import { TracePanelNavigationLayoutDesktop } from "@/src/features/traces/components/TracePanelNavigationLayoutDesktop/TracePanelNavigationLayoutDesktop";
 import { TraceTree } from "@/src/features/traces/components/TraceTree";
 import { TraceTimeline } from "@/src/features/traces/components/TraceTimeline";
 import { useIsMobile } from "@/src/hooks/use-mobile";
