+++ web/src/features/traces/components/TraceLayoutMobile.tsx
