+++ web/src/features/traces/components/TracePanelNavigation.tsx
@@ -17,9 +17,9 @@
 
 import { StringParam, useQueryParam } from "use-query-params";
 import { useSearch } from "@/src/features/traces/contexts/SearchContext";
-import { TraceTree } from "../TraceTree";
-import { TraceSearchList } from "../TraceSearchList";
-import { TraceTimeline } from "../TraceTimeline";
+import { TraceTree } from "./TraceTree";
+import { TraceSearchList } from "./TraceSearchList";
+import { TraceTimeline } from "./TraceTimeline";
 import { useMemo } from "react";
 
 export function TracePanelNavigation() {
