+++ web/src/features/traces/components/TracePanelNavigationHeader/components/TracePanelNavigationButton.tsx
