+++ web/src/features/traces/components/TracePanelNavigationLayoutDesktop/components/TracePanelNavigationHiddenNotice.tsx
